# backend-api/openapi/openapi.yaml
components:
...
  schemas:
...
    DtoHistorySummary:
      properties:
        date:
          format: date-time
          type: string
        idHistory:
          format: int32
          type: integer
        managementId:
          format: int32
          type: integer
        notes:
          type: string
        statusManagementId:
          format: int32
          type: integer
        statusManagementName:
          type: string
      type: object
...
    HistoryRequest:
      properties:
        date:
          format: date-time
          type: string
        managementId:
          format: int32
          type: integer
        managementStatusId:
          format: int32
          type: integer
        notes:
          type: string
      required:
      - managementId
      - managementStatusId
      type: object
...
    PageDtoHistorySummary:
      properties:
        content:
          items:
            $ref: '#/components/schemas/DtoHistorySummary'
          type: array
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          format: int32
          type: integer
        numberOfElements:
          format: int32
          type: integer
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          format: int32
          type: integer
        sort:
          $ref: '#/components/schemas/SortObject'
        totalElements:
          format: int64
          type: integer
        totalPages:
          format: int32
          type: integer
      type: object
...
    PageableObject:
      properties:
        offset:
          format: int64
          type: integer
        pageNumber:
          format: int32
          type: integer
        pageSize:
          format: int32
          type: integer
        paged:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortObject'
        unpaged:
          type: boolean
      type: object
...
    SortObject:
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
      type: object
...
info:
  contact:
    email: info@licensis.com
    name: Licensis
    url: https://www.licensis.com
  description: Sistema de gestión notarial — API REST backend. Todos los recursos siguen el prefijo `/api/v1/`. Las respuestas
    de error siguen el esquema `{error, message, timestamp, path}`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Notaire API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/historial:
    get:
      description: Página de Spring Data (por defecto 20 filas, ordenadas por idHistory ascendente). Usar sort=idHistory,desc
        para ver primero los cambios más recientes. El historial completo de una gestión está en GET /api/v1/historial/gestion/{idManagement}.
      operationId: getAll_15
      parameters:
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          default: 20
          minimum: 1
          type: integer
      - description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort
          criteria are supported.'
        in: query
        name: sort
        required: false
        schema:
          default:
          - idHistory,ASC
          items:
            type: string
          type: array
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageDtoHistorySummary'
          description: OK
      summary: Obtener el historial, una página por vez
      tags:
      - Historial
    post:
      operationId: create_16
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRequest'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Solicitud inválida
        '409':
          content:
            '*/*':
              schema:
                type: object
          description: Conflicto
      summary: Crear nuevo registro de historial
      tags:
      - Historial
  /api/v1/historial/gestion/{idManagement}:
    get:
      operationId: getByManagement
      parameters:
      - in: path
        name: idManagement
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/DtoHistorySummary'
                type: array
          description: OK
      summary: Obtener historial de una gestion (CU13)
      tags:
      - Historial
  /api/v1/historial/{id}:
    delete:
      description: 'El historial es la traza de cambios de estado de la gestión (CU13): solo un administrador puede eliminar
        un registro.'
      operationId: delete_17
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          description: Eliminado
        '403':
          description: Solo un administrador puede eliminar el historial
        '404':
          description: No encontrado
      summary: Eliminar historial (solo administrador)
      tags:
      - Historial
    get:
      operationId: getById_16
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoHistorySummary'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoHistorySummary'
          description: No encontrado
      summary: Obtener historial por ID
      tags:
      - Historial
    put:
      description: 'El historial es la traza de cambios de estado de la gestión (CU13): solo un administrador puede corregir
        un registro.'
      operationId: update_17
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRequest'
        required: true
      responses:
        '200':
          description: OK
        '403':
          description: Solo un administrador puede modificar el historial
        '404':
          description: No encontrado
      summary: Actualizar historial (solo administrador)
      tags:
      - Historial
...
servers:
- description: Desarrollo local
  url: http://localhost:8080
- description: Docker local
  url: http://localhost:8081
...